+++ honey-back/Honeypot/src/main/java/com/honeypot/honeypot/dao/FileSystemInfoDao.java
@@ -16,7 +16,7 @@ public interface FileSystemInfoDao {
      * 根据uniqueid，获取全部文件操作信息
      * @param uniqueId
      * */
-    List<FileSystemInfo> getAllFileOperation(@Param("uniqueId")Integer uniqueId);
+    List<FileSystemInfo> getAllFileOperation(@Param("uniqueId")long uniqueId);
 
     /**
      * 根据查询条件,获取目标文件操作信息
@@ -29,14 +29,14 @@ public interface FileSystemInfoDao {
      * @param time
      * @return
      * */
-    List<FileSystemInfo> getAimFileOperation(@Param("uniqueId")Integer uniqueId,@Param("opType")String opType,@Param("state")String state,@Param("filePath")String filePath,@Param("processPath")String processPath,@Param("processName")String processName,@Param("time")String time);
+    List<FileSystemInfo> getAimFileOperation(@Param("uniqueId")long uniqueId,@Param("opType")String opType,@Param("state")String state,@Param("filePath")String filePath,@Param("processPath")String processPath,@Param("processName")String processName,@Param("time")String time);
 
     /**
      * 根据uniqueid,获取全部文件map操作信息
      * @param uniqueId
      * @return
      * */
-    List<FileSystemInfo> getAllFileMapOperation(@Param("uniqueId")Integer uniqueId);
+    List<FileSystemInfo> getAllFileMapOperation(@Param("uniqueId")long uniqueId);
 
     /**
      * 根据查询条件，获取目标文件map操作详情
@@ -49,5 +49,5 @@ public interface FileSystemInfoDao {
      * @param time
      * @return
      * */
-    List<FileSystemInfo> getAimFileMapOperation(@Param("uniqueId")Integer uniqueId,@Param("opType")String opType,@Param("state")String state,@Param("filePath")String filePath,@Param("processPath")String processPath,@Param("processName")String processName,@Param("time")String time);
+    List<FileSystemInfo> getAimFileMapOperation(@Param("uniqueId")long uniqueId,@Param("opType")String opType,@Param("state")String state,@Param("filePath")String filePath,@Param("processPath")String processPath,@Param("processName")String processName,@Param("time")String time);
 }
